Refactor to standard go layout - separates alert forwarder into separate package to allow for testing/reuse
